fix(conn_packet):optimize define of ll control pdu

+++ common/txType.h
@@ -14,12 +14,12 @@
 #define bool   unsigned char
 
 typedef unsigned long long   _u64;
-typedef unsigned int         _u32;
+typedef unsigned long        _u32;
 typedef unsigned short       _u16;
 typedef unsigned char        _u8;
 
 typedef   signed long long   _s64;
-typedef   signed int         _s32;
+typedef   signed long        _s32;
 typedef   signed short       _s16;
 typedef   signed char        _s8;
